Realtime: enable a playback tracker (#1242)
So far, we've been assuming that audio is played:
- immediately (i.e. with 0 delay/latency)
- at realtime

This causes issues with our interrupt tracking. The model wants to know
how much audio the user has actually heard. For example in a phone call
agent, this wouldn't work (bc theres a delay of a few hundred ms between
model sending audio and the user hearing it). This PR allows you to pass
a playback tracker.






---
[//]: # (BEGIN SAPLING FOOTER)
* #1252
* #1216
* #1243
* __->__ #1242