docs: Add more clarity on multi-agent design patterns (#1650)
This pull request updates the "Agents" documentation page to mention
both manger and handoffs design patterns. I've got a feedback from a
customer that the Agents SDK documentation does not mention the "agents
as tools" design pattern when mentioning handoffs, so he thought this
SDK supports only handoffs. This is a valid feedback, so we can consider
improving the documentation to give more clarity.